refactor(web): 파일명 변경
